Create the style of the header
The background of the header container has been specified, which is a
dark-blue color.

The container was defined as Flex type in order to arrange the header
logo in the upper-left top section, whereas the links were placed in the
upper-right top section. These flex items were aligned to the center
and given space-between.

The format of the text was specifed, i.e, the color, the font-size, and
the font-weight.

In addition, the ul was defined as flex so that the elements
could be placed in a row and be given a gap between them.