Create third container and fix typos
The third container of the main body has been created, which leverages
some parameters of the second container to define the height and
margins.

By arranging the text, a class was added in the html to hook the main
quote. On the other hand, the box of the sentence displaying the
author of the quote was defined as Flex and aligned self to the right
end.