Fix box in container4
The size of the box is bigger by changing the width and height and not
padding, as that did not allow to place the text to the left border
of the box.

On the other hand, the text has the corresponding format and has been
placed in the desired position.